Example update
Changed examples to match function updates.

Updated annotation to reflect the temperature function being the internal temperature of the sensor.

+++ examples/Calibration/Calibration.ino
@@ -73,7 +73,7 @@ void loop()
         Serial.print("CO2 (ppm): ");
         Serial.println(CO2);
 
-        int8_t Temp;    // Buffer for temperature
+        float Temp;    // Buffer for temperature
         Temp = myMHZ19.getTemperature();    // Request Temperature (as Celsius)
 
         Serial.print("Temperature (C): ");
